+++ components/organisms/Forms/FormDashboardSettingsPassword.module.css
@@ -1,0 +1,70 @@
+.form > div:not(:nth-last-child(2)) {
+    border-bottom: 1px solid #C4C4C4;
+}
+
+.form_header {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    margin-bottom: 52px;
+}
+
+.form_header_title {
+    color: var(--font-color-secondary);
+    font-weight: var(--font-weight-bold);
+    font-size: var(--title-size-large);
+}
+
+.wideInput {
+    display: grid;
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+    grid-gap: 128px;
+    align-items: center;
+    gap: 128px;
+    margin-bottom: 32px;
+}
+
+.wideInput_col {
+    display: flex;
+    flex-direction: column;
+}
+
+.wideInput_col_label {
+    color: var(--font-color-secondary);
+    font-weight: var(--font-weight-medium);
+    font-size: var(--title-size-small);
+    margin-bottom: 16px;
+}
+
+.wideInput_col_description {
+    color: var(--font-color-gray1);
+    font-weight: var(--font-weight-regular);
+    font-size: var(--text-size-medium);
+    margin-bottom: 32px;
+}
+
+.form_footer {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+}
+
+@media screen and (max-width: 576px) {
+    .wideInput {
+        grid-template-columns: repeat(1, minmax(0, 1fr));
+        grid-gap: 16px;
+    }
+}
+
+@media screen and (min-width: 576px) and (max-width: 768px) {
+    .wideInput {
+        grid-template-columns: repeat(1, minmax(0, 1fr));
+        grid-gap: 16px;
+    }
+}
+
+@media screen and (min-width: 768px) and (max-width: 1080px) {
+}
+
+@media screen and (min-width: 1080px) {
+}
