+++ components/organisms/Forms/FormForgotPassword.module.css
@@ -1,4 +1,5 @@
 .signUpForm {
+    width: 100%;
     padding-top: 64px;
 }
