+++ components/organisms/Forms/FormSignIn.js
@@ -1,29 +1,36 @@
 import GoogleLogoAsset from '@assets/images/logo_google.png'
 import Button from '@components/atoms/Button'
-import DisplaySuccessOrError from '@components/atoms/FormSuccessOrError'
+import FormSuccessOrError from '@components/atoms/FormSuccessOrError'
 import Input from '@components/atoms/Input'
 import FormHelper from '@helpers/form'
+import TokenHelper from '@helpers/token'
 import ValidationHelper from '@helpers/validation'
 import useServices from '@hooks/useServices'
+import useUser from '@hooks/useUser'
 import { useRouter } from 'next/router'
 import React, { useEffect, useRef } from 'react'
 import { Field, Form } from 'react-final-form'
 import styles from './FormSignIn.module.css'
 
-export default function FormSignUp({}) {
+export default function FormSignIn({}) {
+  const { setUser } = useUser()
   const { auth } = useServices()
   const router = useRouter()
 
-  const refUsername = useRef(null)
-  const refEmail = useRef(null)
+  const refIdentifier = useRef(null)
   const refPassword = useRef(null)
 
   async function handleSubmit(values) {
     try {
-      const res = await auth.api.signUp(values)
+      const res = await auth.api.signIn(values)
 
       if (res.status === 200) {
-        await router.push('/auth/sign-up-email-sent')
+        const { jwt: token, user } = res.data
+
+        TokenHelper.setToken(token)
+        setUser(user)
+
+        await router.push('/')
       } else {
         return ValidationHelper.validateFromBackend(res.data)
       }
@@ -33,49 +40,33 @@ export default function FormSignUp({}) {
   }
 
   useEffect(() => {
-    refUsername.current.focus()
+    refIdentifier.current.focus()
   }, [])
 
   return (
     <Form
-      validate={auth.validations.signUp}
       onSubmit={handleSubmit}
-      render={({ submitError, handleSubmit, values }) => (
+      validate={auth.validations.signIn}
+      render={({ submitError, handleSubmit, values, pristine, submitting }) => (
         <form className={styles.signUpForm}>
-          <DisplaySuccessOrError
+          <FormSuccessOrError
             success={false}
             error={submitError}
             successMessage={''}
           />
 
-          <Field name="username" type="text">
+          <Field name="identifier" type="text">
             {({ input, meta }) => (
               <Input
-                ref={refUsername}
-                icon="person"
-                label="Nom d'utilisateur"
-                placeholder="john.doe"
-                onKeyDown={(e) =>
-                  FormHelper.navigateToNextInput(e, refEmail, 13)
-                }
+                label="Nom d'utilisateur ou email"
+                placeholder="jonhdoe@example.com ou john.doe"
+                ref={refIdentifier}
                 {...input}
                 meta={meta}
-              />
-            )}
-          </Field>
-
-          <Field name="email" type="text">
-            {({ input, meta }) => (
-              <Input
-                ref={refEmail}
-                icon="mail"
-                label="Email"
-                placeholder="john.doe@gmail.com"
+                icon="person"
                 onKeyDown={(e) =>
                   FormHelper.navigateToNextInput(e, refPassword, 13)
                 }
-                {...input}
-                meta={meta}
               />
             )}
           </Field>
@@ -83,15 +74,16 @@ export default function FormSignUp({}) {
           <Field name="password" type="password">
             {({ input, meta }) => (
               <Input
-                ref={refPassword}
-                icon="lock"
+                autocomplete="current-password"
                 label="Mot de passe"
-                placeholder="Choisissez un bon mot de passe"
+                placeholder="Votre mot de passe"
+                ref={refPassword}
+                {...input}
+                meta={meta}
                 onKeyDown={(e) =>
                   FormHelper.withKeyCode(e, 13, () => handleSubmit(values))
                 }
-                {...input}
-                meta={meta}
+                icon="lock"
               />
             )}
           </Field>
@@ -104,11 +96,11 @@ export default function FormSignUp({}) {
               fluid
               onClick={() => handleSubmit(values)}
             >
-              S&apos;inscrire
+              Connexion
             </Button>
 
             <p className={styles.signUpForm__authGroup__separator}>
-              Ou bien se s&apos;inscrire avec
+              Ou bien se connecter avec
             </p>
 
             <Button
@@ -116,7 +108,8 @@ export default function FormSignUp({}) {
               size="large"
               focus="primary"
               fluid
-              onClick={() => console.log('todo: go to goodle')}
+              loading={submitting}
+              onClick={(event) => handleSubmit(event)}
             >
               <img
                 alt="auth-google"
